Update cms-sitename__title to include CMSVersion
First of all, I love your work! I'm using your theme as a replacement for CMS Grouped Menus to reduce the amount of "clutter" in my ModelAdmin column and it works beautifully  :)

When a user or developer is debugging a SilverStripe application, the most common way to identify the version is to log into the CMS and move the mouse over the SilverStripe logo. This is missing in your themed version, so I've ported that feature over from the main CMS project.

Keep up the great work!